requirements.txt
streamlit>=1.35.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
google-generativeai>=0.7.0
rank-bm25>=0.2.2
